+++ CHANGELOG.md
@@ -5,6 +5,13 @@ We are currently working on porting this changelog to the specifications in
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Version 0.2.4] - Unreleased
+
+### Added 
+* Simple mocking for vim args
+* my special "dirty" paragraph formatter that still need to be cleaned up
+
+
 ## [Version 0.2.3] - Unreleased
 
 ### Added